requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
python-multipart
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit
google-genai

